requirements.txt
psycopg2
fastapi
uvicorn
pydantic
python-dotenv